reformatted for optimization, trying to get rid of certain labels.
