+++ src/utils/convertHistoricalData.ts
@@ -1,7 +1,10 @@
 import { HistoricalPriceDataResponse } from "@/lib/types/HistoricalPriceDataResponse";
 
 export const convertHistoricalData = (data: HistoricalPriceDataResponse) => {
-  return data.prices.map((price) => {
+  const filteredData = data.prices.filter((price, index) => {
+    return !(index % 6);
+  });
+  return filteredData.map((price) => {
     return {
       date: new Date(price[0]),
       price: price[1],
